Working GUI
Added this GUI file to get things going.

It is very basic right now.  It opens a GUI and lists the devices on the system.  Right now the RIO laser is hardcoded into the program and is the only device that it will connect to.  When we get more driver files I will update/streamline the code to detect and connect to all hardware