Expanded MongoDB Driver
Includes three classes for accessing database files. First is a "master" which enforces desired database settings before connecting. The second is a "read and write" which connects to the database whithout needing to specify or check any settings. The third is a "read only" connection for observing the database values without editing privileges.

The driver enforces a simple structure for the database. Each named parameter will have its own "database", which contains a permanent, historical record of the parameter, a rolling buffer of the most recent parameter values, and a log for recording notes or comments. Each entry is automatically appended with a UTC timestamp in the datetime.datetime format, and queries to the record and log are selected by timestamp. The log file has an additional "log_level" key to indicate the severity of the message (should integrate this with the stuff in the "logging" folder).

The imagined workflow is to have an instrument control loop read and write to the fast database buffers, and to have a separate "watchdog" loop observe the buffers and decide when to record them into the permanent record.