Better Uptime Bookkeeping
Transfers the last known state from the buffer into the record if it does not already exist there. This removes ambiguities present when estimating the state before and after a record entry with a false initialized flag. Now, if an entry is uninitialized it can be assumed that the time spans before and after that point inherit the compliance of the next initialized record entry.