Uptime Analysis
About as good as I can do with the current data set. I did not preserve the continuity of the database buffer when starting the scripts, so I can't tell from the state information alone what the compliance is before uninitialized states. See the previous commit for the solution going forward.